pyproject.toml
[tool.poetry]
name = "django-ninja-jwt-basic"
version = "0.1.0"
description = ""
authors = ["SpaceShaman <spaceshaman@tuta.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-ninja = "^1.2.1"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
